data-pipeline/cloud_functions/etl_processor/requirements.txt
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
